fixed comments, names and changelog
