+++ src/RedirectUriValidators/RedirectUriValidatorInterface.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * @author      Sebastiano Degan <sebdeg87@gmail.com>
+ * @copyright   Copyright (c) Sebastiano Degan
+ * @license     http://mit-license.org/
+ *
+ * @link        https://github.com/thephpleague/oauth2-server
+ */
+
+namespace League\OAuth2\Server\RedirectUriValidators;
+
+interface RedirectUriValidatorInterface
+{
+    /**
+     * Validates the redirect uri.
+     *
+     * @param string $redirectUri
+     *
+     * @return bool Return true if valid, false otherwise
+     */
+    public function validateRedirectUri($redirectUri);
+}
